Do logfile patching in create_channel helper.
